none

+++ Yukki/Plugins/audioandvideo.py
@@ -62,8 +62,11 @@ async def good(_,CallbackQuery):
 
 @Client.on_callback_query(filters.regex("close"))
 async def closed(_, query: CallbackQuery):
+    a = await _.get_chat_member(query.message.chat.id, query.from_user.id)
+    if not a.can_manage_voice_chats:
+        return await query.answer("💡 only admin can tap this button !", show_alert=True)
     await query.message.delete()
-    await query.answer()  
+    await query.answer()
 
 
 @Client.on_callback_query(filters.regex(pattern=r"down"))
